[FLINK-8038] [table] Clear maps and support cardinality
